Remove code from repositories and move it to routes

+++ packages/api-server/api_server/routes/tasks/tasks.py
@@ -1,3 +1,4 @@
+import uuid
 from datetime import datetime
 from typing import List, Optional, Tuple, cast
 
@@ -6,6 +7,7 @@ from rx import operators as rxops
 from tortoise.exceptions import IntegrityError
 
 from api_server import models as mdl
+from api_server.authenticator import user_dep
 from api_server.dependencies import (
     between_query,
     finish_time_between_query,
@@ -16,6 +18,7 @@ from api_server.dependencies import (
 from api_server.fast_io import FastIORouter, SubscriptionRequest
 from api_server.models import tortoise_models as ttm
 from api_server.models.tortoise_models import TaskState as DbTaskState
+from api_server.models.user import User
 from api_server.repositories import TaskRepository, task_repo_dep
 from api_server.response import RawJSONResponse
 from api_server.rmf_io import task_events, tasks_service
@@ -224,24 +227,56 @@ async def post_undo_skip_phase(
 @router.post("/favorite_task", response_model=ttm.TaskFavoritePydantic)
 async def post_favorite_task(
     request: ttm.TaskFavoritePydantic,
-    task_repo: TaskRepository = Depends(task_repo_dep),
+    user: User = Depends(user_dep),
 ):
     try:
-        await task_repo.save_task_favorite(request)
+        await ttm.TaskFavorite.update_or_create(
+            {
+                "name": request.name,
+                "unix_millis_earliest_start_time": request.unix_millis_earliest_start_time,
+                "priority": request.priority if request.priority else None,
+                "category": request.category,
+                "description": request.description if request.description else None,
+                "user": user.username,
+            },
+            id=request.id if request.id != "" else uuid.uuid4(),
+        )
     except IntegrityError as e:
         raise HTTPException(422, str(e)) from e
 
 
 @router.get("/favorites_tasks", response_model=List[ttm.TaskFavoritePydantic])
 async def get_favorites_tasks(
-    task_repo: TaskRepository = Depends(task_repo_dep),
+    user: User = Depends(user_dep),
 ):
-    return await task_repo.get_all_favorites_tasks()
+    favorites_tasks = await ttm.TaskFavorite.filter(user=user.username)
+    return [
+        ttm.TaskFavoritePydantic(
+            id=favorite_task.id,
+            name=favorite_task.name,
+            unix_millis_earliest_start_time=int(
+                favorite_task.unix_millis_earliest_start_time.strftime("%Y%m%d%H%M%S")
+            )
+            if favorite_task.unix_millis_earliest_start_time
+            else None,
+            priority=favorite_task.priority if favorite_task.priority else None,
+            category=favorite_task.category,
+            description=favorite_task.description
+            if favorite_task.description
+            else None,
+            user=user.username,
+        )
+        for favorite_task in favorites_tasks
+    ]
 
 
 @router.delete("/favorite_task/{favorite_task_id}")
 async def delete_favorite_task(
-    favorite_task_id: str, task_repo: TaskRepository = Depends(task_repo_dep)
+    favorite_task_id: str,
 ):
-    favorite_task = await task_repo.get_favorite_task_by_id(favorite_task_id)
+    favorite_task = await ttm.TaskFavorite.get_or_none(id=favorite_task_id)
+    if favorite_task is None:
+        raise HTTPException(
+            404, f"Favorite task with id {favorite_task_id} does not exists"
+        )
     await favorite_task.delete()
